+++ changelog.md
@@ -1,3 +1,8 @@
+## 3.4.0
+* Add aria-selected attribute, option role, and listbox role for screen readers
+* Add aria-label attribute for screen readers
+* Add keyboard navigation with Tab, Enter, and Space 
+
 ## 3.3.0
 * Update powerbi-visuals-api to 5.4.0
 * Update powerbi-visuals-tools to 4.3.2